requirements.txt
ultralytics>=8.2.0
torch>=2.1.0
torchvision>=0.16.0
opencv-python-headless>=4.9.0
pillow>=10.0.0
pyyaml>=6.0.0
numpy>=1.26.0
matplotlib>=3.8.0
tqdm>=4.66.0
